WIP for #2222
Doesn't work, because userCtx doesn't look to be the same locally.

To test, check out this branch, grunt dev, log in as a local user, and
then in the console:

let DB = angular.element(document.body).injector().get('DB');
DB().post({type: 'a-bad-doc'}).then(console.log).catch(console.error);

This *should* fail locally with a console.error because the doc type
is wrong.

What actually happens is that it incorrectly fails because it thinks
the user is not logged in.
