+++ lib/app.js
@@ -8,6 +8,5 @@ module.exports = {
     fulltext: require('./fulltext'),
     filters: require('./filters'),
     rewrites: require('./rewrites'),
-    views: require('./views'),
-    validate_doc_update: require('./validate_doc_update').validate_doc_update
+    views: require('./views')
 };
