+++ package.json
@@ -32,6 +32,7 @@
     "nools": "^0.4.4",
     "openrosa-xpath-evaluator": "^1.0.0",
     "pouchdb-browser": "^6.0.7",
+    "pouchdb-validation": "^1.2.1",
     "properties": "^1.2.1",
     "select2": "^4.0.3",
     "underscore": "^1.8.3",
